refactor(api): move create certification center usecase
in src/organizational-entities
